feat(ProductDetails) 🎨 rename data folder of data layer to apis folder
